Add basic chat API client
This client implements V1 of the CAPI spec, outlined in the Chat Bot API Alpha v1.pdf document. A sample implementation is included at the bottom of the capi.py file and can be run by calling that file with your API key as the first and only argument.

More information on the API can be found in this thread on the Battle.net forums: https://us.battle.net/forums/en/bnet/topic/20754336617