Add bot instances with commands and config
Config format is JSON. A single file defines all instances and includes user database information. When the bot is started it will check for a config file. If none is found you will be prompted for your API key and a single instance will be created with that key. A config file will be generated during first run and this key will be stored there.

There are 2 commands (ping and whoami) but no interactive way to add users. There is a default group "Admin" which users can be added to (by manually editing the config) for permissions to use the commands.

When a command is registered with a bot instance it defines a permission that is required to use it. Permissions are node-based and  include wildcards (for example the admin group has the 'commands.*' permission).

Added the config file to gitignore.