Create class for main bot instance (manager)
Moves BotInstance class to instance.py. The BnetBot class from bot.py now manages the config loading and management of the instances.

Created PriorityDispatcher, which dispatches events to collections of callbacks based on a requested priority level. The bot instance now handles all of the client events and dispatches them to these handlers. An event callback can veto the event, which stops execution of callbacks for lower priority levels. Chat messages are now printed using this system, and only if a single instance is running.

Adds proper handling of command-line arguments.
 - apikey: adds a new instance with the given API key
 - config: specifies the path to a config file to load
 - debug: prints debugging messages