Improve command handling
Moves internal commands to be defined in commands.py and loaded into each bot instance. Adds several commands: perms, time, whois, ban, unban, kick, designate. Removed their equivalents from the local input processing.

Adds special "%root%" user (a name which can't appear on Battle.net) for executing commands as the bot console.

execute_command() now takes an optional run_as user which will run the command as the specified user. If not given it will default to the user who ran the command (if known). This function will now handle sending responses for unrecognized commands and when a user is in the database but does not have permission for the attempted command.

To add a user to a group, do: /perms group <group> add <user>
To remove a user, do: /perms user <user> remove