Rename temp instances after connection
If an instance created through the --apikey switch, it will be named either "Main" if it is the only instance or "Temp" with a number if there are multiple instances. Once the instance has connected successfully, it will be renamed to match the channel it logs into.

If the provided API key already matches an existing instance, ONLY that instance will be loaded.