Add database write to disk on monitor
The connection checking monitor does more now! It runs every second and will check if the instance's user database has pending writes, and if so writes them. This way changes to the userbase won't be lost if the program is killed (like when you hit the stop button in a debugging environment).